🛠️ [FIX] Using dicts as element IDs now possible
If a dict is used as element ID, it will be correctly converted into a string which is used as react key.
Multiple different elements receive individual keys instead of [object Object] as before. This way the elements can have
individual positions like any other object which does not use a dict as ID.

Signed-off-by: Léon Lazar <leon.lazar@mercedes-benz.com>